Move isolation orchestration to TransformationStep
