+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/transform/DefaultTransformer.java
@@ -20,7 +20,6 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSortedMap;
 import com.google.common.reflect.TypeToken;
 import org.gradle.api.InvalidUserDataException;
-import org.gradle.api.Project;
 import org.gradle.api.artifacts.transform.InputArtifact;
 import org.gradle.api.artifacts.transform.InputArtifactDependencies;
 import org.gradle.api.artifacts.transform.TransformAction;
@@ -30,9 +29,7 @@ import org.gradle.api.file.FileCollection;
 import org.gradle.api.internal.attributes.ImmutableAttributes;
 import org.gradle.api.internal.file.FileCollectionFactory;
 import org.gradle.api.internal.plugins.DslObject;
-import org.gradle.api.internal.project.ProjectStateRegistry;
 import org.gradle.api.internal.tasks.TaskDependencyResolveContext;
-import org.gradle.api.internal.tasks.WorkNodeAction;
 import org.gradle.api.internal.tasks.properties.DefaultParameterValidationContext;
 import org.gradle.api.internal.tasks.properties.FileParameterUtils;
 import org.gradle.api.internal.tasks.properties.InputFilePropertyType;
@@ -86,9 +83,6 @@ public class DefaultTransformer extends AbstractTransformer<TransformAction> {
     private final PropertyWalker parameterPropertyWalker;
     private final boolean requiresDependencies;
     private final InstanceFactory<? extends TransformAction> instanceFactory;
-    private final DomainObjectProjectStateHandler projectStateHandler;
-    private final ProjectStateRegistry.SafeExclusiveLock isolationLock;
-    private final WorkNodeAction isolateAction;
     private final boolean cacheable;
 
     private IsolatableParameters isolatable;
@@ -106,7 +100,6 @@ public class DefaultTransformer extends AbstractTransformer<TransformAction> {
         FileCollectionFactory fileCollectionFactory,
         FileCollectionFingerprinterRegistry fileCollectionFingerprinterRegistry,
         PropertyWalker parameterPropertyWalker,
-        DomainObjectProjectStateHandler projectStateHandler,
         InstantiationScheme actionInstantiationScheme
     ) {
         super(implementationClass, fromAttributes);
@@ -122,20 +115,6 @@ public class DefaultTransformer extends AbstractTransformer<TransformAction> {
         this.instanceFactory = actionInstantiationScheme.forType(implementationClass);
         this.requiresDependencies = instanceFactory.serviceInjectionTriggeredByAnnotation(InputArtifactDependencies.class);
         this.cacheable = cacheable;
-        this.projectStateHandler = projectStateHandler;
-        this.isolationLock = projectStateHandler.newExclusiveOperationLock();
-        this.isolateAction = parameterObject == null ? null : new WorkNodeAction() {
-            @Nullable
-            @Override
-            public Project getProject() {
-                return projectStateHandler.maybeGetOwningProject();
-            }
-
-            @Override
-            public void run() {
-                isolateExclusively();
-            }
-        };
     }
 
     public static void validateInputFileNormalizer(String propertyName, @Nullable Class<? extends FileNormalizer> normalizer, boolean cacheable, ParameterValidationContext parameterValidationContext) {
@@ -157,6 +136,11 @@ public class DefaultTransformer extends AbstractTransformer<TransformAction> {
     @Override
     public Class<? extends FileNormalizer> getInputArtifactDependenciesNormalizer() {
         return dependenciesNormalizer;
+    }
+
+    @Override
+    public boolean isIsolated() {
+        return isolatable != null;
     }
 
     public boolean requiresDependencies() {
@@ -183,9 +167,6 @@ public class DefaultTransformer extends AbstractTransformer<TransformAction> {
 
     @Override
     public void visitDependencies(TaskDependencyResolveContext context) {
-        if (isolateAction != null) {
-            context.add(isolateAction);
-        }
         if (parameterObject != null) {
             parameterPropertyWalker.visitProperties(parameterObject, ParameterValidationContext.NOOP, new PropertyVisitor.Adapter() {
                 @Override
@@ -198,26 +179,11 @@ public class DefaultTransformer extends AbstractTransformer<TransformAction> {
 
     @Override
     public void isolateParameters() {
-        if (isolatable == null) {
-            if (!projectStateHandler.hasMutableProjectState()) {
-                projectStateHandler.withLenientState(this::isolateExclusively);
-            } else {
-                isolateExclusively();
-            }
+        try {
+            isolatable = doIsolateParameters();
+        } catch (Exception e) {
+            throw new VariantTransformConfigurationException(String.format("Cannot isolate parameters %s of artifact transform %s", parameterObject, ModelType.of(getImplementationClass()).getDisplayName()), e);
         }
-    }
-
-    private void isolateExclusively() {
-        isolationLock.withLock(() -> {
-            if (isolatable != null) {
-                return;
-            }
-            try {
-                isolatable = doIsolateParameters();
-            } catch (Exception e) {
-                throw new VariantTransformConfigurationException(String.format("Cannot isolate parameters %s of artifact transform %s", parameterObject, ModelType.of(getImplementationClass()).getDisplayName()), e);
-            }
-        });
     }
 
     protected IsolatableParameters doIsolateParameters() {
